Generate guess for maker mode
Created an ai method to generate computer guesses
for when user plays in Codemaker mode. The method
takes in a zipped form of an array that shows the
previous guess with either B (black), W (white),
or blank. It stores B or W numbers in an array and
generates another guess using any B and W numbers
along with random other numbers from RANGE.